+++ src/c/overflow.c
@@ -1,0 +1,8 @@
+
+#include <stdio.h>
+int main()
+{
+	unsigned char umax = 254;
+	umax += 5;
+	printf("%i", umax);
+}