+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2020-03-02
++ Have to introduce binary addition before introducing method to convert negative decimal to twos complement
++ Split binary addition and subtraction into two seperate sections
++ Added a C - examples folder
+
 ## 2020-03-01
 + Split Negative Binary Numbers from Binary Addition/Subtraction into a different section
     + Added going from bit patterns to signed and unsigned numbers